requirements.txt
langchain==0.3.21
langchain-core==0.3.79
langchain-community==0.3.20
langchain-google-genai==2.0.10
langchain-pinecone==0.2.12
pinecone-client==5.0.1
google-generativeai==0.8.3
sentence-transformers==3.3.1
torch==2.5.1
torchvision==0.20.1
pypdf==5.1.0
google-search-results==2.4.2
Pillow==11.0.0
streamlit==1.40.2
requests==2.32.3
python-dotenv==1.0.1
loguru
google-genai